ufo destroy: improve error handling for in progress state

+++ lib/ufo/destroy.rb
@@ -8,8 +8,16 @@ module Ufo
 
       stack = find_stack(@stack_name)
       unless stack
-        puts "Stack #{@stack_name} does not exist."
+        puts "Stack #{@stack_name.colorize(:green)} does not exist."
         exit
+      end
+
+      if stack.stack_status =~ /_IN_PROGRESS$/
+        puts "Cannot destroy service #{@pretty_service_name.colorize(:green)}"
+        puts "Cannot delete stack #{@stack_name.colorize(:green)} in this state: #{stack.stack_status.colorize(:green)}"
+        puts "If the stack is taking a long time, you can cancel the current operation with:"
+        puts "  ufo cancel #{@service}"
+        return
       end
 
       cloudformation.delete_stack(stack_name: @stack_name)
